ks_bridge: Update set return type of ksb_debug_show() to int
This function is passed as an argument to single_open() which expects it to have
a return type of int, not ssize_t.

+++ drivers/usb/misc/ks_bridge.c
@@ -952,7 +952,7 @@ static struct usb_driver ksb_usb_driver = {
 	.supports_autosuspend = 1,
 };
 
-static ssize_t ksb_debug_show(struct seq_file *s, void *unused)
+static int ksb_debug_show(struct seq_file *s, void *unused)
 {
 	unsigned long		flags;
 	struct ks_bridge	*ksb = s->private;
